# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Pedidos",
    "description": "Documentación de la API para gestión de usuarios, productos y pedidos",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4001/api"
    }
  ],
  "tags": [
    { "name": "Users" },
    { "name": "Orders" },
    { "name": "Products" },
    { "name": "Categories" },
    { "name": "Payments" }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Obtener todos los usuarios",
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/User" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Crear usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Obtener usuario por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "Actualizar usuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado"
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Eliminar usuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado"
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Obtener productos",
        "responses": {
          "200": {
            "description": "Lista de productos"
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Crear producto",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado"
          }
        }
      }
    },
    "/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Obtener pedidos",
        "responses": {
          "200": {
            "description": "Lista de pedidos"
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Crear pedido",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido creado"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Obtener categorías",
        "responses": {
          "200": {
            "description": "Lista de categorías"
          }
        }
      }
    },
    "/payments": {
      "get": {
        "tags": ["Payments"],
        "summary": "Obtener pagos",
        "responses": {
          "200": {
            "description": "Lista de pagos"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "string" },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserInput": {
        "type": "object",
        "required": ["name", "email", "phone", "password"],
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "password": { "type": "string" },
          "role": { "type": "string" }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "price": { "type": "number" },
          "description": { "type": "string" },
          "categoryId": { "type": "integer" }
        }
      },
      "ProductInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "price": { "type": "number" },
          "description": { "type": "string" },
          "category_id": { "type": "integer" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "delivery_address": { "type": "string" },
          "delivery_reference": { "type": "string" },
          "total": { "type": "number" },
          "status": { "type": "string" },
          "payment_method": { "type": "string" }
        }
      },
      "OrderInput": {
        "type": "object",
        "properties": {
          "userId": { "type": "integer" },
          "deliveryAddress": { "type": "string" },
          "deliveryReference": { "type": "string" },
          "total": { "type": "number" },
          "paymentMethod": { "type": "string" },
          "paymentStatus":{"type":"string"}
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" }
        }
      },
      "Payment": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "orderId": { "type": "integer" },
          "transactionId":{ "type": "number" },
          "provider":{ "type": "string" },
          "amount": { "type": "number" },
          "status": { "type": "string" }
        }
      }
    }
  }
}